Move the chroot call to probe_worker
We have discovered a problem when testing offline mode in the symlink
probe. When the probe main function needs to be executed multiple times,
function `probe_common_main`, where the chroot is entered, is called
only once, but function `probe_worker`, which leaves the chroot, is called
many times. This leads to a situation that the first OVAL test is
evaluated in chroot, but all further tests have retrieved data from the
host. Therefore the collected data were wrong. We try to address this
problem by moving the code.
