+++ src/OVAL/probes/probe/probe_main.c
@@ -130,22 +130,6 @@ static int probe_opthandler_rcache(int option, int op, va_list args)
 	return (0);
 }
 
-// Dummy pthread routine
-static void * dummy_routine(void *dummy_param)
-{
-	return NULL;
-}
-
-static void preload_libraries_before_chroot()
-{
-	// Force to load dynamic libraries used by pthread_cancel
-	pthread_t t;
-	if (pthread_create(&t, NULL, dummy_routine, NULL))
-		fail(errno, "pthread_create(probe_preload)", __LINE__ - 1);
-	pthread_cancel(t);
-	pthread_join(t, NULL);
-}
-
 static void probe_common_main_cleanup(void *arg)
 {
 	dD("probe_common_main_cleanup started");
@@ -183,7 +167,6 @@ void *probe_common_main(void *arg)
 {
 	pthread_attr_t th_attr;
 	probe_t        probe;
-	char *rootdir = NULL;
 	struct probe_common_main_argument *probe_argument = (struct probe_common_main_argument *) arg;
 	sch_queuedata_t *data = probe_argument->queuedata;
 	oval_subtype_t subtype = probe_argument->subtype;
@@ -250,52 +233,6 @@ void *probe_common_main(void *arg)
 	OSCAP_GSYM(probe_optdef) = probe.option;
 	OSCAP_GSYM(probe_optdef_count) = probe.optcnt;
 
-#ifndef OS_WINDOWS
-	probe_offline_mode_function_t offline_mode_function = probe_table_get_offline_mode_function(probe.subtype);
-	if (offline_mode_function != NULL) {
-		probe.supported_offline_mode = offline_mode_function();
-	}
-
-	/*
-	 * Setup offline mode(s)
-	 */
-	rootdir = getenv("OSCAP_PROBE_ROOT");
-	if ((rootdir != NULL) && (strlen(rootdir) > 0)) {
-		probe.offline_mode = true;
-
-		preload_libraries_before_chroot(); // todo - maybe useless for own mode
-
-		if (probe.supported_offline_mode & PROBE_OFFLINE_OWN) {
-			dI("Swiching probe to PROBE_OFFLINE_OWN mode.");
-			probe.selected_offline_mode = PROBE_OFFLINE_OWN;
-
-		} else if (probe.supported_offline_mode & PROBE_OFFLINE_CHROOT) {
-			probe.real_root_fd = open("/", O_RDONLY);
-			probe.real_cwd_fd = open(".", O_RDONLY);
-			if (chdir(rootdir) != 0) {
-				fail(errno, "chdir", __LINE__ -1);
-			}
-
-			if (chroot(rootdir) != 0) {
-				fail(errno, "chroot", __LINE__ - 1);
-			}
-			/* NOTE: We're running in a different root directory.
-			 * Unless /proc, /sys are somehow emulated for the new
-			 * environment, they are not relevant and so are other
-			 * runtime only things (e.g. getenv, uname, ...).
-			 * Switch to offline mode. We may add a separate
-			 * mechanism to control this behaviour in the future.
-			 */
-			dI("Swiching probe to PROBE_OFFLINE_CHROOT mode.");
-			probe.selected_offline_mode = PROBE_OFFLINE_CHROOT;
-		}
-	}
-
-	if (getenv("OSCAP_PROBE_RPMDB_PATH") != NULL) {
-		dI("Swiching probe to PROBE_OFFLINE_RPMDB mode.");
-		probe.selected_offline_mode = PROBE_OFFLINE_RPMDB;
-	}
-#endif
 
 	/*
 	 * Create input handler (detached)
